# backend/openapi/uniflow-api.yaml
openapi: 3.0.3
info:
  title: Uniflow API
  version: 1.0.0
  description: >
    Specification OpenAPI des endpoints actuellement implementes dans le backend Uniflow.
servers:
  - url: http://localhost:8080
    description: Serveur local Spring Boot
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Inscrire un utilisateur standard
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Utilisateur cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authentifier un compte
      operationId: loginAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Authentification reussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/events:
    get:
      tags: [Events]
      summary: Lister les evenements par pole
      operationId: getEventsByPole
      parameters:
        - in: query
          name: pole
          required: true
          schema:
            type: string
          example: Informatique
      responses:
        '200':
          description: Liste des evenements du pole demande
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
  /api/events/create:
    post:
      tags: [Events]
      summary: Creer un evenement
      operationId: createEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Evenement cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/events/{id}:
    get:
      tags: [Events]
      summary: Recuperer un evenement par identifiant
      operationId: getEventById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Evenement trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDTO'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Events]
      summary: Modifier un evenement
      operationId: updateEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: Evenement modifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Events]
      summary: Supprimer un evenement
      operationId: deleteEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Evenement supprime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/events/orga/{organisationId}:
    get:
      tags: [Events]
      summary: Lister les evenements d'une organisation
      operationId: getEventsByOrganisation
      parameters:
        - in: path
          name: organisationId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Liste des evenements de l'organisation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
  /api/events/favori/{userId}:
    get:
      tags: [Events]
      summary: Lister les evenements favoris d'un utilisateur
      operationId: getFavoriteEvents
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Liste des favoris
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/events/favori/{userId}/{eventId}:
    post:
      tags: [Events]
      summary: Ajouter un evenement aux favoris
      operationId: addFavoriteEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Favori ajoute
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Events]
      summary: Retirer un evenement des favoris
      operationId: removeFavoriteEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Favori supprime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/inscriptions/event/{eventId}:
    get:
      tags: [Inscriptions]
      summary: Lister les participants d'un evenement
      operationId: getEventParticipants
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Liste des participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipantDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
    UserId:
      in: path
      name: userId
      required: true
      schema:
        type: integer
        format: int64
    EventId:
      in: path
      name: eventId
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Donnees invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentification requise ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Acces interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        password:
          type: string
          example: secret123
    RegisterRequest:
      type: object
      required: [nom, prenom, email, password]
      properties:
        nom:
          type: string
          example: Dupont
        prenom:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        password:
          type: string
          example: secret123
    AuthResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        roles:
          type: array
          items:
            type: string
            enum: [ADMIN, ORGANISATEUR, UTILISATEUR]
          example: [UTILISATEUR, ORGANISATEUR]
        displayName:
          type: string
          example: Alice Dupont
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        message:
          type: string
          example: Connexion reussie.
        token:
          type: string
          description: JWT a envoyer dans le header Authorization Bearer.
        organisations:
          type: array
          items:
            type: string
          example: [Club Informatique]
...
    Event:
      type: object
      required: [title, organisation]
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        title:
          type: string
          example: Conference Intelligence Artificielle
        description:
          type: string
          example: Decouvrez les dernieres avancees en intelligence artificielle.
        address:
          type: string
          example: Amphitheatre A - Campus Lyon 1
        beginDate:
          type: string
          format: date-time
          example: 2026-04-15T09:00:00
        endDate:
          type: string
          format: date-time
          example: 2026-04-15T17:00:00
        capacity:
          type: integer
          nullable: true
          example: 150
        status:
          type: string
          enum: [PUBLISHED, CLOSED, CANCELED]
          example: PUBLISHED
        price:
          type: number
          format: double
          example: 25.0
        organisation:
          type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              example: 1
        pole:
          type: object
          nullable: true
          properties:
            id:
              type: integer
              format: int64
              example: 1
    EventDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Conference Intelligence Artificielle
        description:
          type: string
          example: Decouvrez les dernieres avancees en intelligence artificielle.
        address:
          type: string
          example: Amphitheatre Themis - Campus Lyon 1
        beginDate:
          type: string
          format: date-time
          example: 2026-04-15T09:00:00
        endDate:
          type: string
          format: date-time
          example: 2026-04-15T17:00:00
        capacity:
          type: integer
          nullable: true
          example: 150
        status:
          type: string
          enum: [PUBLISHED, CLOSED, CANCELED]
          example: PUBLISHED
        price:
          type: number
          format: double
          example: 25.0
        organisationId:
          type: integer
          format: int64
          nullable: true
          example: 1
        organisationName:
          type: string
          nullable: true
          example: Club Informatique
        poleId:
          type: integer
          format: int64
          nullable: true
          example: 1
        poleName:
          type: string
          nullable: true
          example: Informatique
...
    ParticipantDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nom:
          type: string
          example: Dupont
        prenom:
          type: string
          example: Alice
        email:
          type: string
          format: email
          example: alice@univ-lyon1.fr
        eventId:
          type: integer
          format: int64
          example: 12
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation effectuee avec succes
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Identifiants invalides.